refactor: move code out of main package
