+++ internal/config/config.go
@@ -1,4 +1,4 @@
-package operator
+package config
 
 import (
 	"time"
