+++ internal/controller/stas/scan_job_controller.go
@@ -23,18 +23,18 @@ import (
 	"sigs.k8s.io/json"
 
 	stasv1alpha1 "github.com/statnett/image-scanner-operator/api/stas/v1alpha1"
+	"github.com/statnett/image-scanner-operator/internal/config"
 	"github.com/statnett/image-scanner-operator/internal/controller"
 	staserrors "github.com/statnett/image-scanner-operator/internal/errors"
 	"github.com/statnett/image-scanner-operator/internal/pod"
 	"github.com/statnett/image-scanner-operator/internal/trivy"
-	"github.com/statnett/image-scanner-operator/pkg/operator"
 )
 
 // ScanJobReconciler reconciles a finished image scan Job object.
 type ScanJobReconciler struct {
 	client.Client
 	Scheme *runtime.Scheme
-	operator.Config
+	config.Config
 	pod.LogsReader
 }
